internal/ast: simplify precedence table
It was supposed to use the same rules as Go, but I actually implemented the C rules
instead. I find the Go table easier to reason about since most operators just associate
left to right. There are effectively two levels you need to be aware of:
*/%&<< binds stronger than +-|^.

With the C precedence table, << is weaker than + for example, and & is weaker still.
It's useful and probably well intended but few people actually want to memorize this.
In practice expressions will be parenthesized anyway for readability.